ERP-26_Product :: Manage Status flag 0- Pending(Initial) 1- Publish 2- UnPublish 3- Delete

+++ MidCapERP.Dto/Product/ProductRequestDto.cs
@@ -62,7 +62,7 @@ namespace MidCapERP.Dto.Product
         public string? QRImage { get; set; }
 
         public int TenantId { get; set; }
-        public int? Status { get; set; }
+        public byte? Status { get; set; }
         public int CreatedBy { get; set; }
         public string CreatedByName { get; set; }
         public DateTime CreatedDate { get; set; }
